Remove [Nuget MyGet] color tests
