+++ services/nuget/nuget.tester.js
@@ -4,15 +4,34 @@ import {
   isVPlusDottedVersionNClauses,
   isVPlusDottedVersionNClausesWithOptionalSuffix,
 } from '../test-validators.js'
-import {
-  queryIndex,
-  nuGetV3VersionJsonFirstCharZero,
-  nuGetV3VersionJsonFirstCharNotZero,
-  nuGetV3VersionJsonBuildMetadataWithDash,
-} from '../nuget-fixtures.js'
 import { invalidJSON } from '../response-fixtures.js'
 
 export const t = new ServiceTester({ id: 'nuget', title: 'NuGet' })
+
+const queryIndex = JSON.stringify({
+  resources: [
+    {
+      '@id': 'https://api-v2v3search-0.nuget.org/query',
+      '@type': 'SearchQueryService',
+    },
+  ],
+})
+
+const nuGetV3VersionJsonBuildMetadataWithDash = JSON.stringify({
+  data: [
+    {
+      totalDownloads: 0,
+      versions: [
+        {
+          version: '1.16.0+388',
+        },
+        {
+          version: '1.17.0+1b81349-429',
+        },
+      ],
+    },
+  ],
+})
 
 // downloads
 
@@ -50,59 +69,6 @@ t.create('version (valid)')
     message: isVPlusDottedVersionNClauses,
   })
 
-t.create('version (orange badge)')
-  .get('/v/Microsoft.AspNetCore.Mvc.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get(
-        '/query?q=packageid%3Amicrosoft.aspnetcore.mvc&prerelease=true&semVerLevel=2',
-      )
-      .reply(200, nuGetV3VersionJsonFirstCharZero),
-  )
-  .expectBadge({
-    label: 'nuget',
-    message: 'v0.35',
-    color: 'orange',
-  })
-
-t.create('version (blue badge)')
-  .get('/v/Microsoft.AspNetCore.Mvc.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get(
-        '/query?q=packageid%3Amicrosoft.aspnetcore.mvc&prerelease=true&semVerLevel=2',
-      )
-      .reply(200, nuGetV3VersionJsonFirstCharNotZero),
-  )
-  .expectBadge({
-    label: 'nuget',
-    message: 'v1.2.7',
-    color: 'blue',
-  })
-
-// https://github.com/badges/shields/issues/4219
-t.create('version (build metadata with -)')
-  .get('/v/MongoFramework.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get('/query?q=packageid%3Amongoframework&prerelease=true&semVerLevel=2')
-      .reply(200, nuGetV3VersionJsonBuildMetadataWithDash),
-  )
-  .expectBadge({
-    label: 'nuget',
-    message: 'v1.17.0',
-    color: 'blue',
-  })
-
 t.create('version (not found)')
   .get('/v/not-a-real-package.json')
   .expectBadge({ label: 'nuget', message: 'package not found' })
@@ -121,6 +87,23 @@ t.create('version (unexpected second response)')
   )
   .expectBadge({ label: 'nuget', message: 'unparseable json response' })
 
+// https://github.com/badges/shields/issues/4219
+t.create('version (build metadata with -)')
+  .get('/v/MongoFramework.json')
+  .intercept(nock =>
+    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
+  )
+  .intercept(nock =>
+    nock('https://api-v2v3search-0.nuget.org')
+      .get('/query?q=packageid%3Amongoframework&prerelease=true&semVerLevel=2')
+      .reply(200, nuGetV3VersionJsonBuildMetadataWithDash),
+  )
+  .expectBadge({
+    label: 'nuget',
+    message: 'v1.17.0',
+    color: 'blue',
+  })
+
 // version (pre)
 
 t.create('version (pre) (valid)')
@@ -130,56 +113,6 @@ t.create('version (pre) (valid)')
     message: isVPlusDottedVersionNClausesWithOptionalSuffix,
   })
 
-t.create('version (pre) (orange badge)')
-  .get('/vpre/Microsoft.AspNetCore.Mvc.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get(
-        '/query?q=packageid%3Amicrosoft.aspnetcore.mvc&prerelease=true&semVerLevel=2',
-      )
-      .reply(200, nuGetV3VersionJsonFirstCharZero),
-  )
-  .expectBadge({
-    label: 'nuget',
-    message: 'v0.35',
-    color: 'orange',
-  })
-
-t.create('version (pre) (blue badge)')
-  .get('/vpre/Microsoft.AspNetCore.Mvc.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get(
-        '/query?q=packageid%3Amicrosoft.aspnetcore.mvc&prerelease=true&semVerLevel=2',
-      )
-      .reply(200, nuGetV3VersionJsonFirstCharNotZero),
-  )
-  .expectBadge({
-    label: 'nuget',
-    message: 'v1.2.7',
-    color: 'blue',
-  })
-
 t.create('version (pre) (not found)')
   .get('/vpre/not-a-real-package.json')
   .expectBadge({ label: 'nuget', message: 'package not found' })
-
-t.create('version (pre) (unexpected second response)')
-  .get('/vpre/Microsoft.AspNetCore.Mvc.json')
-  .intercept(nock =>
-    nock('https://api.nuget.org').get('/v3/index.json').reply(200, queryIndex),
-  )
-  .intercept(nock =>
-    nock('https://api-v2v3search-0.nuget.org')
-      .get(
-        '/query?q=packageid%3Amicrosoft.aspnetcore.mvc&prerelease=true&semVerLevel=2',
-      )
-      .reply(invalidJSON),
-  )
-  .expectBadge({ label: 'nuget', message: 'unparseable json response' })
